requirements.txt
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
numpy>=1.19.0
scipy>=1.5.0
scikit-learn>=0.23.1
pandas>=1.0.5
Pillow==7.2.0
streamlit==0.62.1